Add price, tvl tables and enable filtering on components

+++ tycho-types/src/dto.rs
@@ -511,7 +511,7 @@ pub struct ProtocolComponentsRequestBody {
 #[derive(Serialize, Deserialize, Default, Debug, IntoParams)]
 pub struct ProtocolComponentRequestParameters {
     #[param(default = 0)]
-    pub tvl_gt: Option<u64>,
+    pub tvl_gt: Option<f64>,
 }
 
 impl ProtocolComponentsRequestBody {
